moved logic with dynamically disabling buttons to filter.vue in order to relieve low level component
