+++ common/types/filters.ts
@@ -1,0 +1,4 @@
+export type TCalendarDisabledButtons = {
+    today: boolean,
+    tomorrow: boolean
+}
