+++ frontend/components/common/ui/DateTimepicker.vue
@@ -2,16 +2,11 @@
 import VueDatePicker, { type DatePickerInstance } from '@vuepic/vue-datepicker';
 import '@vuepic/vue-datepicker/dist/main.css';
 import dayjs from 'dayjs';
+import type { TCalendarDisabledButtons } from '../../../../common/types/filters'
 import type { PropType } from 'vue';
 import type { Time } from '../../../utils/dates';
 
 // https://vue3datepicker.com/props/modes/
-
-type TDisabled = {
-	today: boolean,
-	tomorrow: boolean
-}
-
 const props = defineProps({
 	className: {
 		type: String,
@@ -64,6 +59,15 @@ const props = defineProps({
 	appearance: {
 		type: String as PropType<'no-border' | null>,
 		default: null
+	},
+	disabledButtons: {
+		type: Object as PropType<TCalendarDisabledButtons>,
+		default: () => {
+			return {
+				today: false,
+				tomorrow: false
+			}
+		}
 	}
 });
 
@@ -93,29 +97,6 @@ const onRemove = () => {
 
 const today = new Date()
 const tomorrow = new Date(new Date().setDate(today.getDate() + 1))
-
-const isDisabled = computed((): TDisabled => {
-	const isEndDate = props.name === 'endDate'
-	// обе кнопки !disabled
-	const result = {
-		today: false,
-		tomorrow: false
-	}
-	
-	// кнопка сегодня disabled
-	if (isEndDate && today.getTime() <= new Date(props.minDate).getTime()) {
-		result.today = true
-	}
-
-	// обе кнопки disabled
-	if (isEndDate && tomorrow.getTime() <= new Date(props.minDate).getTime()) {
-		result.today = true
-		result.tomorrow = true
-	}
-
-	return result
-})
-
 </script>
 
 <template>
@@ -167,7 +148,7 @@ const isDisabled = computed((): TDisabled => {
 				<CommonButton
 					:button-text="$t('dates.filterDay.today')"
 					button-kind="dark"
-					:is-disabled="isDisabled.today"
+					:is-disabled="disabledButtons.today"
 					@click="() => {
 						$emit('update:model-value', today)
 						datepicker.closeMenu()
@@ -176,7 +157,7 @@ const isDisabled = computed((): TDisabled => {
 				<CommonButton
 					:button-text="$t('dates.filterDay.tomorrow')"
 					button-kind="dark"
-					:is-disabled="isDisabled.tomorrow"
+					:is-disabled="disabledButtons.tomorrow"
 					@click="() => {
 						$emit('update:model-value', tomorrow)
 						datepicker.closeMenu()
